git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@14430 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/org/jdownloader/settings/GeneralSettings.java
@@ -161,5 +161,5 @@ public interface GeneralSettings extends ConfigInterface {
     @Description("Penaltytime before a retry if JDownloader lost connection")
     long getWaittimeOnConnectionLoss();
 
-    void setWaittimeOnConnectionLoss(int milliseconds);
+    void setWaittimeOnConnectionLoss(long milliseconds);
 }
